- Project settings applying required upgrades due to upgrade to Xcode 7.2 - Applying upgrades to ExerciseTableViewCell, RootViewController, ExerciseListTableViewController, EditRowListTableViewController, EditUserListTableViewController

+++ Classes/EditRowListTableViewController.m
@@ -248,7 +248,7 @@
 
 - (void)tableView:(UITableView *)tableView didSelectRowAtIndexPath:(NSIndexPath *)indexPath {
     // Navigation logic may go here. Create and push another view controller.
-	NSLog(@"selected row %i", indexPath.row);
+	NSLog(@"selected row %li", (long)indexPath.row);
 }
